+++ src/test/java/fr/insee/queen/api/campaign/controller/CampaignControllerTest.java
@@ -1,6 +1,5 @@
 package fr.insee.queen.api.campaign.controller;
 
-import fr.insee.queen.api.campaign.controller.dto.output.CampaignSummaryDto;
 import fr.insee.queen.api.campaign.service.dummy.CampaignFakeService;
 import fr.insee.queen.api.campaign.service.exception.CampaignDeletionException;
 import fr.insee.queen.api.pilotage.controller.dummy.PilotageFakeComponent;
@@ -9,8 +8,6 @@ import fr.insee.queen.api.utils.dummy.AuthenticationFakeHelper;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-
-import java.util.List;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
@@ -52,12 +49,5 @@ class CampaignControllerTest {
                 .isInstanceOf(CampaignDeletionException.class);
     }
 
-    @Test
-    @DisplayName("On retrieving interviewer campaigns, all interviewer campaigns are retrieved")
-    void testGetInterviewerCampaigns01() {
-        List<CampaignSummaryDto> campaigns = campaignController.getInterviewerCampaignList();
-        assertThat(pilotageComponent.wentThroughInterviewerCampaigns()).isTrue();
-        assertThat(campaigns).hasSize(2);
-        assertThat(campaigns.get(0).id()).isEqualTo(PilotageFakeComponent.CAMPAIGN1_ID);
-    }
+
 }
